requirements.txt
spotipy>=2.23.0
python-dotenv>=1.0.0
streamlit>=1.30.0
requests>=2.28.0
